requirements.txt
matplotlib==3.8.0
numpy==1.26.3
pandas==2.2.0
Pillow==10.0.1
Pillow==10.2.0
plotly==5.9.0
scikit_learn==1.3.0
seaborn==0.13.1
tensorflow==2.10.0
